Revert "Create article"
