+++ .storybook/addons.js
@@ -1,2 +1,2 @@
-// import '@storybook/addon-actions/register';
-// import '@storybook/addon-links/register';
+import '@storybook/addon-actions/register';
+import '@storybook/addon-links/register';
