+++ src/App.css
@@ -14,7 +14,6 @@
 button {
   border: none;
   outline: 0;
-  width: 85px;
 }
 
 .spacer {
