+++ src/components/ArticleForm.js
@@ -1,187 +1,56 @@
-import React, { Component } from 'react'
+import React from 'react'
 import store from '../store'
-// import ArticlePreview, { demoArticle } from './ArticlePreview.js'
-import Modale from './Modale.js'
-import './css/ArticleForm.css'
 
-const freshArticle = {
-  title: '',
-  shortDescription: '',
-  section: '',
-  headerImage: '',
-  tags:'',
-  hasStar:false,
-  content: []
-}
+const ArticleForm = () => {
+  const state = store.getState()
 
-const H2 = ({ element, ...rest }) => {
-  return (
-    <label>Ajouter un titre de paragraphe :
-    <input type="text" value={element.value} {...rest} />
-  </label>
-  )
-}
+  const handleChange = event => {
+    console.log({ target: event.target, [event.target.name]: event.target.value })
 
-const P = ({ element, ...rest }) => {
-  return (
-    <label>Ajouter un paragraphe :
-    <textarea type="text" value={element.value} {...rest} />
-    </label>
-  )
-}
-
-const Blockquote = ({ element, ...rest }) => {
-  return (
-    <label>Ajouter une citation :
-    <input type="text" value={element.value} {...rest} />
-    </label>
-  )
-}
-
-const Imgs = ({ element, ...rest }) => {
-  return (
-    <label>Ajouter des images :
-    <input type="text" value={element.value} {...rest} />
-    </label>
-  )
-}
-
-const toInput = {
-  h2: (props) => <H2 {...props} />,
-  p: (props) => <P {...props} />,
-  blockquote: (props) => <Blockquote {...props} />,
-  imgs: (props) => <Imgs {...props} />,
-}
-
-const Element = (props) => toInput[props.element.type](props)
-
-class ArticleForm extends Component {
-  state = {
-    article: freshArticle // demoArticle
-  }
-
-  handleChange = event => {
-    const key = event.target.name
-
-    let article = {}
-
-    if (key.startsWith('content')) {
-      const index = key.split('-')[1]
-
-      const content = [ ...this.state.article.content ]
-      content[index].value = event.target.value // issue: either text or url or urls
-
-      article = {
-        ...this.state.article,
-        content: content
-      }
-    } else {
-      article = {
-        ...this.state.article,
-        [key]: event.target.value
-      }
-    }
-
-    this.setState({ article })
-  }
-
-  handleSubmit = event => {
-    event.preventDefault()
-
-    const article = {
-
-      title: this.state.article.title,
-      shortDescription: this.state.article.shortDescription,
-      headerImage: this.state.article.headerImage,
-      section: this.state.article.section,
-      hasStar: this.state.article.hasStar,
-      tags: this.state.article.tags,
-      content: JSON.stringify(this.state.article.content)
-    }
-    console.log(this.state.article)
-    console.log(article)
-
-    fetch('http://localhost:3456/articles', {
-      method: 'post',
-      body: JSON.stringify(article),
-      headers: {
-        'Content-Type': 'application/json'
-      }
+    store.dispatch({
+      type: 'FORM_INPUT_CHANGED',
+      formId: 'ADD_ARTICLE',
+      inputName: event.target.name,
+      inputValue: event.target.value
     })
   }
 
-  addInput = type => {
-    const basicContentElement = {
-      h2: { "type": "h2", "value": "" },
-      p: { "type": "p", "value": "" },
-      blockquote: { "type": "blockquote", "value": "" },
-      imgs: { "type": "imgs", "value": "" },
-    }
+  const handleSubmit = event => {
+    event.preventDefault()
 
-    const article = {
-      ...this.state.article,
-      content: [
-        ...this.state.article.content,
-        basicContentElement[type]
-      ]
-    }
+    const state = store.getState()
+    const newArticle = state.articleform.forms['ADD_ARTICLE']
 
-    this.setState({ article })
+    console.log('SUBMIT', newArticle)
+    console.log('SUBMIT', JSON.stringify(newArticle))
+
+    fetch('http://localhost:3456/articles', { method: 'post',
+      body: JSON.stringify(newArticle),
+      headers: {
+        'Content-Type': 'application/json'
+      } })
   }
 
-  render () {
-    console.log(this.state.hasStar)
+  const form = state.articleform.forms['ADD_ARTICLE']
 
-    const article = this.state.article
-
-    const buttons = [ 'h2' , 'p', 'blockquote', 'imgs' ]
-      .map((type, i) => <button key={i} onClick={() => this.addInput(type)}>{type}</button>)
-
-    const dynamicInputs = article.content
-      .map((element, i) => <Element key={i} name={`content-${i}`} element={element} onChange={this.handleChange} />)
-
-    return (
-      <div className="box">
-        <div className="item-left">
-          <div style={{marginTop: '15px'}}>
-            <form onSubmit={this.handleSubmit}>
-              <label>Titre de l'article :<br/>
-                <input type="text" name="title" value={article.title} onChange={this.handleChange} />
-              </label>
-              <label>Description de l'article :<br/>
-              <textarea type="text" name="shortDescription" value={article.shortDescription} onChange={this.handleChange} />
-              </label>
-              <label>URL de l'image de couverture:<br/>
-                <input type="text" name="headerImage" value={article.headerImage} onChange={this.handleChange} />
-              </label>
-              <label>Tags de l'article :<br/>
-                <input type="text" name="tags" value={article.tags} onChange={this.handleChange} />
-              </label>
-              <label>
-              <input style={{width: '20px', marginRight: '5px'}} type="checkbox" name="hasStar" onClick={() => this.setState({hasStar: !this.state.hasStar}) } />
-              Mettre à la une</label>
-              <div>
-                <select name="section" value={article.section} onChange={this.handleChange}>
-                  <option value="Choose">Choisissez la section</option>
-                  <option value="Lab">Lab</option>
-                  <option value="Projet">Projet</option>
-                </select>
-              </div>
-
-              <div className="addModule yellow">Ajouter un module :</div>
-              {dynamicInputs}
-              <div id="buttons" style={{ backgroundColor: '#fbd052', marginBottom: '20px' }}>{buttons}</div>
-              <input type="submit" value="Submit" />
-            </form>
-          </div>
-        </div>
-        <div className="item-right">
-          <Modale article={article} />
-        </div>
-      </div>
-    )
-  }
-
+  return (
+    <form onSubmit={handleSubmit}>
+      <h3>Titre:</h3>
+      <input type="text" name="title" value={form.title} onChange={handleChange} />
+      <h3>Description:</h3>
+      <textarea type="text" name="shortDescription" value={form.shortDescription} onChange={handleChange}>
+      </textarea>
+      <h3>URL de l'image de couverture:</h3>
+      <textarea type="text" name="content" value={form.headerImage} onChange={handleChange}>
+      </textarea>
+      <select name="section" value={form.section} onChange={handleChange}>
+        <option value="Choose">Choose</option>
+        <option value="Lab">Lab</option>
+        <option value="Projet">Projet</option>
+      </select>
+      <input type="submit" value="Submit" />
+    </form>
+  )
 }
 
 export default ArticleForm
