+++ src/components/Modale.js
@@ -2,32 +2,15 @@ import React from 'react'
 import './css/Modale.css'
 import { Link } from '@reach/router'
 
-const toHTML = {
-  h2: ({ value }) => <h4>{value}</h4>,
-  p: ({ value }) => <p>{value}</p>,
-  blockquote: ({ value }) => <blockquote className='quote'>{value}</blockquote>,
-  // img: ({ url }) => <img src={url} />,
-  imgs: ({ value }) => value.split(',').map((url, i) => <img key={i} src={url} />),
-}
-
-const Element = ({ element }) => toHTML[element.type](element)
-
-
 const Modale = ({ article, displayModale }) => {
   // const reversedPathname = window.location.pathname.split('').reverse()
   // const index = reversedPathname.indexOf('/')
   // const parentContextPath = reversedPathname.slice(index + 1).reverse().join('')
-  if (typeof article.content === "string")
-    JSON.parse(article.content)
-
-  const content = article.content
-    .map((element, i) => <Element key={i} element={element} />)
-
 
   const parentContextPath = window.location.pathname.replace(/\d+$/, '')
 
   return (
-    <div className="ModaleBlock">
+    <div className="ModaleBlock" style={{ display: `${displayModale}` }}>
       <Link className="closeModaleBtn" to={parentContextPath}><div className="closeModaleBtn">✕</div></Link>
       <div className="ModalePic" style={{ background: `center / cover no-repeat url(${article.headerImage})` }}>
         <div className="ModaleHeader FilterBlack" style={{ padding: '0.1rem 0.75rem' }}>
@@ -42,7 +25,9 @@ const Modale = ({ article, displayModale }) => {
           </p>
         </div>
       </div>
-      {content}
+      <p>
+        {article.content}
+      </p>
     </div>
   )
 }
